+++ example/mcp_driver/mcp_mock_server.dart
@@ -20,24 +20,18 @@ void main() async {
     inputSchemaProperties: {},
     callback: ({args, extra}) async {
       int count = mockUtil.count();
-      return CallToolResult(
-        content: [TextContent(text: 'count: $count')],
-      );
+      return CallToolResult(content: [TextContent(text: 'count: $count')],);
     },
   );
 
   server.tool(
     'create',
     description: 'Create a text in storage',
-    inputSchemaProperties: {
-      'text': {'type': 'string'},
-    },
+    inputSchemaProperties: {'text': {'type': 'string'},},
     callback: ({args, extra}) async {
       String text = args!['text'] as String;
       int id = mockUtil.create(text);
-      return CallToolResult(
-        content: [TextContent(text: 'id: $id')],
-      );
+      return CallToolResult(content: [TextContent(text: 'id: $id')],);
     },
   );
 
@@ -47,41 +41,30 @@ void main() async {
     callback: ({args, extra}) async {
       int id = args!["id"] as int;
       String text = mockUtil.read(id);
-      return CallToolResult(
-        content: [TextContent(text: 'text: $text')],
-      );
+      return CallToolResult(content: [TextContent(text: 'text: $text')],);
     },
   );
 
   server.tool(
     'update',
     description: 'Update a text in storage by id',
-    inputSchemaProperties: {
-      'id': {'type': 'integer'},
-      'text': {'type': 'string'}
-    },
+    inputSchemaProperties: {'id': {'type': 'integer'}, 'text': {'type': 'string'}},
     callback: ({args, extra}) async {
       int id = args!['id'] as int;
       String text = args['text'] as String;
       mockUtil.update(id, text);
-      return CallToolResult(
-        content: [TextContent(text: '{"result": "Update successfully."}')],
-      );
+      return CallToolResult(content: [TextContent(text: '{"result": "Update successfully."}')],);
     },
   );
 
   server.tool(
     'delete',
     description: 'Delete a text in storage',
-    inputSchemaProperties: {
-      'id': {'type': 'integer'},
-    },
+    inputSchemaProperties: {'id': {'type': 'integer'},},
     callback: ({args, extra}) async {
       int id = args!['id'] as int;
       mockUtil.delete(id);
-      return CallToolResult(
-        content: [TextContent(text: '{"result": "Delete successfully."}')],
-      );
+      return CallToolResult(content: [TextContent(text: '{"result": "Delete successfully."}')],);
     },
   );